Lincando elementos paginas

+++ app/Http/Controllers/User/VanController.php
@@ -56,6 +56,7 @@ class VanController extends Controller
         $van->model = $request->model;
         $van->plate = $request->plate;
         $van->seats = $request->seats;
+        $van->comment = $request->comment;
 
         $van->save();
 
@@ -95,11 +96,13 @@ class VanController extends Controller
         }
 
 
-        foreach ($request->van_user_photo as $eachPhoto) {
-            $van_user_photo = new VanUserPhoto();
-            $van_user_photo->van_id = $van->id;
-            $van_user_photo->user_photo_id = $eachPhoto;
-            $van_user_photo->save();
+        if (isset($request->van_user_photo)) {
+            foreach ($request->van_user_photo as $eachPhoto) {
+                $van_user_photo = new VanUserPhoto();
+                $van_user_photo->van_id = $van->id;
+                $van_user_photo->user_photo_id = $eachPhoto;
+                $van_user_photo->save();
+            }
         }
 
         return view('user.van');
@@ -133,8 +136,7 @@ class VanController extends Controller
         $photos_selected = $van->van_photos($id);
 
         $array_photos_selected = array();
-        foreach($photos_selected as $eachPhotoSelected)
-        {
+        foreach ($photos_selected as $eachPhotoSelected) {
             $array_photos_selected[] = $eachPhotoSelected->user_photo_id;
         }
 
@@ -175,6 +177,7 @@ class VanController extends Controller
         $van->model = $request->model;
         $van->plate = $request->plate;
         $van->seats = $request->seats;
+        $van->comment = $request->comment;
 
         $van->save();
 
@@ -220,13 +223,14 @@ class VanController extends Controller
             VanTrack::where("van_id", "=", $id)->delete();
         }
 
-
-        VanUserPhoto::where("van_id", "=", $van->id)->delete();
-        foreach ($request->van_user_photo as $eachPhoto) {
-            $van_user_photo = new VanUserPhoto();
-            $van_user_photo->van_id = $van->id;
-            $van_user_photo->user_photo_id = $eachPhoto;
-            $van_user_photo->save();
+        if (isset($request->van_user_photo)) {
+            VanUserPhoto::where("van_id", "=", $van->id)->delete();
+            foreach ($request->van_user_photo as $eachPhoto) {
+                $van_user_photo = new VanUserPhoto();
+                $van_user_photo->van_id = $van->id;
+                $van_user_photo->user_photo_id = $eachPhoto;
+                $van_user_photo->save();
+            }
         }
 
         return redirect('user/van');
